requirements.txt
azure-devops==7.1.0b4
msrest==0.7.1
openai>=1.14.0
anthropic>=0.42.0
python-dotenv==1.0.1
rich==13.7.1
requests>=2.31.0
